Fix to avoid gcc warnings
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/branches/double-coords-migration@328 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ FermaNext/src/TrussUnit/TrussUnitWindow.cpp
@@ -1028,8 +1028,8 @@ void TrussUnitWindow::drawTrussArea ( ren_dynarow& baseRend,
     double areaWidInPix = p2.y() - p1.y();
     double scaleFactorXInPix = areaLenInPix / 5;
     double scaleFactorYInPix = areaWidInPix / 5;
-    int scaleFactorXInAbs = getTrussAreaSize().height() / 5;
-    int scaleFactorYInAbs = getTrussAreaSize().width() / 5;
+    int scaleFactorXInAbs = int(getTrussAreaSize().height() / 5);
+    int scaleFactorYInAbs = int(getTrussAreaSize().width() / 5);
 
     p2.setX ( p1.x() - scalePieceLength );
     QString str;
